perubahan dikit

+++ src/byte_sector_conversion.rs
@@ -1,7 +1,7 @@
 pub fn sector_to_byte(input: String) -> String {
     let string_data = input.replace("s", "");
 
-    let conversion_data = string_data.trim().parse::<u64>().unwrap();
+    let conversion_data = string_data.trim().parse::<u128>().unwrap();
 
     let conversion_data = conversion_data * 512;
